refactor(desktop): extract flatpak logic into dedicated module
